requirements.txt
pip
rich
openai
tiktoken
pygame-ce
python-dotenv
openai-whisper
SpeechRecognition